Refactor bl12381
